refactor(utilities): move component and enhance layout
