+++ src/routes/utilities.lazy.tsx
@@ -1,80 +1,19 @@
-import { Page } from "@/components/common/Page";
-import { Button } from "@/components/ui/button";
-import { windowOpen } from "@/utils/windowOpen";
 import { createLazyFileRoute } from "@tanstack/react-router";
-
-const utilities = [
-  {
-    label: "Alchemix Stats",
-    author: "Barree",
-    url: "https://alchemix-stats.com/",
-    image: "alchemix_stats.png",
-  },
-  {
-    label: "Self-Repaying-ENS",
-    author: "Wary",
-    url: "https://ens.alchemix.fi/",
-    image: "srens.png",
-  },
-  {
-    label: "AMO Harvest Tool",
-    author: "Amrit & Theblockdudes",
-    url: "https://alchemix-multisig.netlify.app/",
-    image: "amo_harvester.png", 
-  },
-  {
-    label: "Contract Params Dashboard",
-    author: "Build3rsLabs",
-    url: "https://alchemix-dashboard-frontend-lovat.vercel.app/",
-    image: "contract_params_dashboard.png", 
-  },
-];
+import { Page } from "@/components/common/Page";
+import { Utilities } from "@/components/utilities/Utilities";
 
 export const Route = createLazyFileRoute("/utilities")({
-  component: Utilities,
+  component: UtilitiesRoute,
 });
 
-function Utilities() {
+function UtilitiesRoute() {
   return (
     <Page
       title="Utilities"
       description="A collection of useful tools"
       iconUri="/images/icons/utilities_thin.svg"
     >
-      <p className="mb-6 text-center text-xs opacity-50">
-        These tools are developed and maintained by our awesome community
-        members.
-      </p>
-      <div className="flex w-full flex-row flex-wrap gap-4">
-        {utilities.map((utility) => (
-          <div
-            key={utility.label}
-            className="w-full md:w-1/4 rounded border border-grey10inverse bg-grey15inverse dark:border-grey10 dark:bg-grey15"
-          >
-            <p className="border-b border-grey10inverse px-4 py-2 dark:border-grey10">
-              {utility.label}
-            </p>
-            <div
-              className="h-48 bg-cover bg-center"
-              style={{
-                backgroundImage: `url('../images/screenshots/${utility.image}')`,
-              }}
-            ></div>
-            <div className="flex flex-row items-center justify-between border-t border-grey10inverse px-4 py-2 dark:border-grey10">
-              <div className="flex flex-row space-x-4">
-                <p>{utility.author}</p>
-              </div>
-              <Button
-                variant="ghost"
-                className="h-8 w-20 sm:w-32 md:w-40 lg:w-max border border-bronze1 text-base"
-                onClick={() => windowOpen(utility.url)}
-              >
-                Open
-              </Button>
-            </div>
-          </div>
-        ))}
-      </div>
+      <Utilities />
     </Page>
   );
 }
